Create DOC ASSISTANT .
The Doc Assistant  is an application that assist both the health workers and patients in the quest of improving the medical sector. As well all know, the he health sector is one of the most important aspect of out country and the world as a whole, it has to be at it's best. 
The problems that we mostly faced in the health sector are as follows; long queues, increased number of diseases outbreak due to lack of information and ignorance on the patient's side and also Negative attitude from the health workers to the patients.
All these problems can be solved easily by our application which is the Doc Assistance. Our application will be a link between the health workers which are doctors and patients to interact with each other without the patient be visiting the hospital or clinic. The Doc Assist application will be user-friendly meaning that when you have no idea on how to use the application, it will be your guide. It will help you learn how to use it by giving you options on which language you desire to listen to. It will be Reliable, it won't entirely reply on the Internet for it to function only, it can also be used as a reminder when the user needs to take medication or when it is time to exercise. It will be fast and efficient; incase of any emergency, it will be able to provide a proper first aid guide as you wait for the health workers to reach your location. In cases where you are beaten by a snake and you have no knowledge on how to go about it, the app will assist you with first aid guide to buy you some time to visit the nearest clinic. It will have a GPS Service that health workers will be using to locate patients who are unable to go to the hospital by themselves. It will be mostly used for emergency like pregnancy or any other emergency you might have. It will also be helpful to keep disease outbreak in control, when they notice that three or five people in the same area have the same disease,they will be able to send people to put it in control. The All Assistant will be included to help people who lack Self confidence or those who have no one to share they fears, worries and happiness with. Our Application will be there to comfort them and act as there met or and friend. This feature will create a sense of friend whose loyal,trustworthy,helpful,loving and kind.  Through this application a user can freely and peacefully express themselves. It wil also be directed to a doctor in case the matter is serious and more sensitive that will require a doctor to advise the user.